moved all methods from homework to a new structur
