+++ page-objects/homework.js
@@ -7,13 +7,9 @@ const dateTimeHelpers = require('../runtime/helpers/dateTimeHelpers.js');
 const elementHelpers = require('../runtime/helpers/elementHelpers.js');
 const courseData = require('../shared-objects/courseData');
 const Login = require('../shared-objects/loginData');
-const copyCourse = require('../page-objects/copyCourse');
 const firstLogin = require('../shared_steps/firstLogin.js');
 const loginPage = require('../page-objects/pages/generalPagesBeforeLogin/LoginPage.js');
-const navigationTopPage = require('../page-objects/pages/NavigationTopPage.js');
 const startPage = require('../page-objects/pages/generalPagesBeforeLogin/StartPageBeforeLogin.js');
-const { logout } = require('./pages/NavigationTopPage.js');
-const addCoursePage = require("../page-objects/pages/coursePages/CRSSAddCoursePage");
 const courseListPage = require("../page-objects/pages/coursePages/CRSSCourseListPage");
 // TODO: choose course, SORT
 
@@ -21,34 +17,7 @@ const click = async (selector) => (await driver.$(selector)).click();
 
 module.exports = {
 	// add homework related functions (as a teacher)
-	clickCreateNewTaskInTheCourse: async function (coursename) {
-		await courseListPage.clickOnCourseInSection(coursename, courseListPage.section.activeCourses);
-		let homeworktab = await driver.$('.tabs button[data-testid="hometasks"]');
-		await homeworktab.click();
-		await waitHelpers.waitAndClick(courseData.elem.addHomeworkBtn);
-	},
-	setPrivate: async function () {
-		await waitHelpers.waitAndClick(courseData.elem.checkbox);
-	},
-	addBasicHometask: async function (coursename, taskname) {
-		await this.clickCreateNewTaskInTheCourse(coursename);
-		let nameSelector = await driver.$(courseData.elem.homeworkName);
-		await nameSelector.setValue(taskname);
-		await waitHelpers.waitAndClick(courseData.elem.teamworkActivate);
-		await this.setAccomplishTime();
-		await this.setHometaskText();
-		await waitHelpers.waitAndClick(courseData.elem.submitAddHomeworkBtn);
-	},
-	addPrivateHometask: async function (coursename, taskname) {
-		await this.clickCreateNewTaskInTheCourse(coursename);
-		let nameSelector = await driver.$(courseData.elem.homeworkName);
-		await nameSelector.setValue(taskname);
-		await waitHelpers.waitAndClick(courseData.elem.teamworkActivate);
-		await this.setAccomplishTime();
-		await this.setHometaskText();
-		await this.setPrivate();
-		await waitHelpers.waitAndClick(courseData.elem.submitAddHomeworkBtn);
-	},
+
 	setHometaskText: async function () {
 		await driver.pause(global.SHORT_WAIT_MILLIS);
 		const editorContent = await driver.$('.ck-content');
@@ -69,15 +38,7 @@ module.exports = {
 		await selectorToBeLoaded.waitForExist(2000);
 	},
 
-	gotoTasks: async function () {
-		await elementHelpers.loadPage(courseData.urlHomework, 20);
-	},
-
-	gotoTasksTab: async function () {
-		let hometasksTab = await driver.$('button[data-testid="hometasks"]');
-		await hometasksTab.click();
-		await driver.pause(1000);
-	},
+	
 
 	sortHometasks: async function () {
 		let sortBtn = await driver.$('#filter > div > div.md-chip.md-theme-default.md-deletable.md-clickable > div');
@@ -107,10 +68,7 @@ module.exports = {
 			}
 		}
 	},
-	areThereAnyTasks: async function () {
-		let elementWithTasks = await driver.$$('.col-xl-12');
-		return elementWithTasks.length > 0 ? true : false;
-	},
+	
 	chooseTaskAmongAllTasks: async function (taskname) {
 		let taskindex = await this.returnTaskIndex(taskname);
 		if (taskindex != false) {
@@ -136,88 +94,16 @@ module.exports = {
 		await expect(taskName).to.equal(taskname);
 	},
 
-	teacherLogsIn: async function () {
-		await this.userLogsOut();
-		await startPage.clickLoginBtn();
-		await loginPage.performLogin(Login.defaultTeacherUsername, Login.defaultTeacherpassword);
-	},
-	goToTasksOfTheCourse: async function (coursename) {
-		await courseListPage.goToCourses();
-		await courseListPage.clickOnCourseInSection(coursename, courseListPage.section.activeCourses);
-		await this.gotoTasksTab();
-	},
-	studentLogsInAndGoesToTasksOfTheCourse: async function (username, password, coursename) {
-		await this.userLogsOut();
-		await firstLogin.pupilLogin(username, password);
-		await firstLogin.firstLoginPupilFullAge(username, password);
-		await this.goToTasksOfTheCourse(coursename);
-	},
-	privateTaskVerify: async function () {
-		let areThereAnyTasks = await this.areThereAnyTasks();
-		if (areThereAnyTasks == true) {
-			let taskNames = await Promise.all(
-				(await driver.$$('#homeworks > ol > div > li > a')).map(async (element) => await element.getText()),
-			);
-			await expect(taskNames).not.to.include(taskname);
-			return;
-		}
-		await expect(areThereAnyTasks).to.be.false;
-	},
+	
+	
 
-	userLogsOut: async function () {
-		await elementHelpers.loadPage(courseData.urlLogout, 20);
-	},
+	
 	// student helpers
-	userFindsTheTask: async function (taskname) {
-		let areThereAnyTasks = await driver.$$('#homeworks > ol > div > li');
-		await expect(areThereAnyTasks.length).not.to.equal(0);
-		for (var i = 1; i <= areThereAnyTasks.length; i++) {
-			let taskSelector = await driver.$('#homeworks > ol > div > li:nth-child(' + i + ') .h5.title');
-			let tasknameOnPage = await taskSelector.getText();
-			if (tasknameOnPage == taskname) {
-				await taskSelector.click();
-				await driver.pause(1000);
-			}
-		}
-	},
-	switchToSubmissionTab: async function () {
-		let submissionTab = "#submission-tab-link";
-		await waitHelpers.waitAndClick(submissionTab);
-	},
-	submitSolutionForTheHometask: async function () {
-		await driver.pause(global.SHORT_WAIT_MILLIS);
-		const textField = await driver.$('.ck-content');
-		const assignmentText = 'here is some text which I want to submit';
-		textField.setValue(assignmentText);
-		const container = await driver.$('#submission');
-		const submitBtn = await container.$('button[type="submit"]');
-		await waitHelpers.waitAndClick('.ckeditor-submit')
-		await driver.pause(1500);
-	},
-
-	studentEditsTextHomeworkAndSubmits: async function () {
-		await this.switchToSubmissionTab();
-		await this.submitSolutionForTheHometask();
-	},
+	
+	
 
 	// teacher helpers
-	hasTheStudentSubmittedTheTask: async function (studentname) {
-		let submissionTab = "#submissions-tab-link";
-		await waitHelpers.waitAndClick(submissionTab);
-		let submitted_by_box = await driver.$('#submissions .groupNames > span');
-		let submitted_by_name = await submitted_by_box.getText();
-		await expect(submitted_by_name).to.contain(studentname);
-	},
-
-	teacherLogsInAndCanSeeTheTextSubmission: async function (coursename, taskname, studentname) {
-		await this.teacherLogsIn();
-		await firstLogin.firstLoginTeacher();
-		await courseListPage.goToCourses();
-		await courseListPage.clickOnCourseInSection(coursename, courseListPage.section.activeCourses);
-		await this.gotoTasksTab();
-		await this.userFindsTheTask(taskname);
-		await this.hasTheStudentSubmittedTheTask(studentname);
-	},
+	
 	evaluateSubmission: async function () {
 		let submittedTasks = await driver.$('.usersubmission');
 		await submittedTasks.click();
@@ -231,16 +117,8 @@ module.exports = {
 		await textField.setValue(comment);
 	},
 
-	uploadAHomework: async function () {
-		//making the upload-element visible to selenium
-		change_visibility =
-			'$x("//*[@id="main-content"]/div/section[1]/div/div/div[1]/input").css("visibility,"visible");';
-		change_display = '$x("//*[@id="main-content"]/div/section[1]/div/div/div[1]/input").css("display,"block");';
-		await driver.execute_script(change_visibility);
-		await driver.execute_script(change_display);
-
-		const filePath = path.join(__dirname, '../shared-objects/fileUpldFolder/upload.txt');
-		await driver.$x(courseData.uploadBtn).send_keys(filePath);
+	gotoTasks: async function () {
+		await elementHelpers.loadPage(courseData.urlHomework, 20);
 	},
 
 	submitHomework: async function (taskName, student) {
@@ -250,29 +128,7 @@ module.exports = {
 		await this.submitSolutionForTheHometask();
 	},
 
-	submitFileFeedback: async function (taskName, file) {
-		// 	back to teacher
-		await this.userLogsOut();
-		await this.teacherLogsIn();
-		// grade the submission
-		await this.gotoTasks();
-		await waitHelpers.waitAndClick(`*=${taskName}`);
-
-		await this.teacherShowGradeTabForFirstSubmission();
-
-		// upload the file
-		await driver.execute(function () {
-			// Need to make the input visible, otherwise the webdriver can not upload any files
-			document.querySelector('input[type=file][class=dz-hidden-input]').style = {};
-		});
-
-		const remoteFilePath = await driver.uploadFile(file.path);
-		await (await driver.$('input[type=file][class=dz-hidden-input]')).setValue(remoteFilePath);
-		await driver.pause(3000);
-
-		// The upload causes a page reload, which causes the current tab to change.
-		await (await driver.$('.tab-content.section-homeworksubmissions.active')).waitForDisplayed();
-	},
+	
 
 	testFileUploadSuccess: async function (taskName, file, student) {
 		// navigate to grade tab
@@ -309,11 +165,7 @@ module.exports = {
 		expect(studentFileUrl.pathname).to.equal(fileUrl.pathname);
 	},
 
-	teacherShowGradeTabForFirstSubmission: async function () {
-		await waitHelpers.waitAndClick('#submissions-tab-link');
-		await click('tbody.usersubmission');
-		await click('a*=Bewertung');
-	},
+	
 
 	canSeeFile: async function (file) {
 		const gradeFilesList = await driver.$('.list-group-files');
