+++ page-objects/pages/HMWRKHomeworkListPage.js
@@ -116,5 +116,35 @@ module.exports = {
 		const filePath = path.join(__dirname, '../shared-objects/fileUpldFolder/upload.txt');
 		await driver.$x(courseData.uploadBtn).send_keys(filePath);
 	},
+
+	areThereAnyTasks: async function () {
+		let elementWithTasks = await driver.$$('.col-xl-12');
+		return elementWithTasks.length > 0 ? true : false;
+	},
+
+	privateTaskVerify: async function () {
+		let areThereAnyTasks = await this.areThereAnyTasks();
+		if (areThereAnyTasks == true) {
+			let taskNames = await Promise.all(
+				(await driver.$$('#homeworks > ol > div > li > a')).map(async (element) => await element.getText()),
+			);
+			await expect(taskNames).not.to.include(taskname);
+			return;
+		}
+		await expect(areThereAnyTasks).to.be.false;
+	},
+
+	userFindsTheTask: async function (taskname) {
+		let areThereAnyTasks = await driver.$$('#homeworks > ol > div > li');
+		await expect(areThereAnyTasks.length).not.to.equal(0);
+		for (var i = 1; i <= areThereAnyTasks.length; i++) {
+			let taskSelector = await driver.$('#homeworks > ol > div > li:nth-child(' + i + ') .h5.title');
+			let tasknameOnPage = await taskSelector.getText();
+			if (tasknameOnPage == taskname) {
+				await taskSelector.click();
+				await driver.pause(1000);
+			}
+		}
+	},
 };
